+++ tps/serv-impressao/src/servidor_usuario.c
@@ -2,7 +2,7 @@
 #include <string.h>
 #include "servidor_usuario.h"
 
-TUsuario* TUsuario_Criar(const char* Nome, const size_t Prioridade)
+TUsuario* TUsuario_Criar(const char* Nome, const int Prioridade)
 {
 	TUsuario* NovoUsuario;
